\encoding{latin1}
\name{mail}
\alias{mail}
\docType{package}
\title{
Sending Email Notifications from R
}
\description{
Easy to use package for sending email notifications with status information from R.
}
\details{
\tabular{ll}{
Package: \tab mail - Rpackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-31\cr
License: \tab GPL version 2 or later\cr
Maintainer: \tab Dr. Lin Himmelmann <Rmail@linhi.com>\cr
URL: \tab  www.linhi.com\cr
}
}
